services/ocr_service/requirements.txt
rapidocr-onnxruntime>=1.2.0
Pillow>=10.2.0
kafka-python>=2.0.2
tornado>=6.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
